Multi-node data preparation added.

+++ data_preparation/preprocess.py
@@ -8,6 +8,7 @@ def main(cfg):
     bignlp_path = cfg.bignlp_path
     data_cfg = cfg.data_preparation
     data_dir = cfg.data_dir
+    rm_extracted = data_cfg.rm_extracted
     assert data_dir is not None, "data_dir must be a valid path"
 
     # Vocab
@@ -20,46 +21,28 @@ def main(cfg):
     assert merges_dir is not None, "merges_save_dir must be a valid path."
     merges_path = os.path.join(bignlp_path, merges_dir, "merges.txt")
 
-    file_number = int(os.environ.get("SLURM_ARRAY_TASK_ID"))
-    extracted_path = os.path.join(data_dir, f"{file_number:02d}.jsonl")
-    code_path = "/opt/bignlp/NeMo/scripts/nlp_language_modeling/preprocess_data_for_megatron.py"
-    output_prefix = os.path.join(data_dir, f"my-gpt3_{file_number:02d}")
-
-    flags = (
-        f"--input {extracted_path} "
-        f"--output-prefix {output_prefix} "
-        f"--vocab {vocab_path} "
-        f"--merge-file {merges_path} "
-        f"--dataset-impl mmap "
-        f"--tokenizer-library megatron "
-        f"--tokenizer-type GPT2BPETokenizer "
-        f"--workers $SLURM_CPUS_ON_NODE "
-        f"--append-eod "
+    # This compile doesn't seem to do anything. It compiles
+    # "helpers.cpython-38-x86_64-linux-gnu.so", but since that file already
+    # exists, it doesn't do anything. Force make via: touch helpers.cpp
+    megatron_dir = '/opt/bignlp/NeMo/nemo/collections/nlp/data/language_modeling/megatron'
+    compiled_helpers_lib = os.path.join(megatron_dir, 'compiled_helpers_lib')
+    compilecmd = (
+            f'cd /opt/bignlp/NeMo; git rev-parse HEAD; '
+            f'cd {megatron_dir}; '
+            f'touch helpers.cpp; make;'
     )
-    cmd = f'cd /opt/bignlp/NeMo; git rev-parse HEAD; cd /opt/bignlp/NeMo/nemo/collections/nlp/data/language_modeling/megatron; make; export PYTHONPATH="/opt/bignlp/NeMo/.:$PYTHONPATH"; export TRANSFORMERS_CACHE="/temp_root/.cache/"; CUDA_VISIBLE_DEVICES=0 python3 {code_path} {flags}'
-    os.system(f"{cmd}")
-    os.remove(extracted_path)
 
+    code_path = "/opt/bignlp/NeMo/scripts/nlp_language_modeling/preprocess_data_for_megatron.py"
+    runcmd = (
+        f'cd {megatron_dir}; '
+        f'export PYTHONPATH="/opt/bignlp/NeMo/.:$PYTHONPATH"; '
+        f'export TRANSFORMERS_CACHE="/temp_root/.cache/"; '
+        f'CUDA_VISIBLE_DEVICES=0,4,2,6,1,5,3,7 python3 {code_path} ' + '{flags}'
+    )
 
-def preprocess_bcp(cfg, file_numbers):
-    bignlp_path = cfg.bignlp_path
-    data_cfg = cfg.data_preparation
-    data_dir = cfg.data_dir
-    assert data_dir is not None, "data_dir must be a valid path"
-
-    # Vocab
-    vocab_dir = data_cfg.vocab_save_dir
-    assert vocab_dir is not None, "vocab_save_dir must be a valid path."
-    vocab_path = os.path.join(bignlp_path, vocab_dir, "vocab.json")
-
-    # Merges
-    merges_dir = data_cfg.merges_save_dir
-    assert merges_dir is not None, "merges_save_dir must be a valid path."
-    merges_path = os.path.join(bignlp_path, merges_dir, "merges.txt")
-
-    for file_number in file_numbers:
+    if cfg.cluster_type == "bcm":
+        file_number = int(os.environ.get("SLURM_ARRAY_TASK_ID"))
         extracted_path = os.path.join(data_dir, f"{file_number:02d}.jsonl")
-        code_path = "/opt/bignlp/NeMo/scripts/nlp_language_modeling/preprocess_data_for_megatron.py"
         output_prefix = os.path.join(data_dir, f"my-gpt3_{file_number:02d}")
 
         flags = (
@@ -70,13 +53,52 @@ def preprocess_bcp(cfg, file_numbers):
             f"--dataset-impl mmap "
             f"--tokenizer-library megatron "
             f"--tokenizer-type GPT2BPETokenizer "
-            f"--workers 50 "
+            f"--workers $SLURM_CPUS_ON_NODE "
             f"--append-eod "
         )
-        cmd = f'cd /opt/bignlp/NeMo; git rev-parse HEAD; cd /opt/bignlp/NeMo/nemo/collections/nlp/data/language_modeling/megatron; make; export PYTHONPATH="/opt/bignlp/NeMo/.:$PYTHONPATH"; export TRANSFORMERS_CACHE="/temp_root/.cache/"; CUDA_VISIBLE_DEVICES=0 python3 {code_path} {flags}'
-        print(f"Preprocessing file {extracted_path}")
-        os.system(f"{cmd}")
-        os.remove(extracted_path)
+        os.system(compilecmd)
+        os.system(runcmd.format(flags=flags))
+        if rm_extracted:
+            os.remove(extracted_path)
+    elif cfg.cluster_type == "bcp":
+        file_numbers = data_cfg.file_numbers
+        files_list = utils.convert_file_numbers(file_numbers)
+        # Assumes launched via mpirun:
+        #   mpirun -N <nnodes> -npernode 1 ...
+        wrank = int(os.environ.get("OMPI_COMM_WORLD_RANK", 0))
+        wsize = int(os.environ.get("OMPI_COMM_WORLD_SIZE", 0))
+        lrank = int(os.environ.get("OMPI_COMM_WORLD_LOCAL_RANK", 0))
+
+        if lrank == 0:
+            # Compile once per node. Should be one container instance per node.
+            os.system(compilecmd)
+            os.system(f"touch {compiled_helpers_lib}")
+        else:
+            while not os.path.exists(compiled_helpers_lib):
+                sleep(1)
+
+        files_list_groups = utils.split_list(files_list, wsize)
+        files_to_preproc = files_list_groups[wrank]
+        ncpus = psutil.cpu_count(logical=False)
+        for file_number in files_to_preproc:
+            extracted_path = os.path.join(data_dir, f"{file_number:02d}.jsonl")
+            output_prefix = os.path.join(data_dir, f"my-gpt3_{file_number:02d}")
+
+            flags = (
+                f"--input {extracted_path} "
+                f"--output-prefix {output_prefix} "
+                f"--vocab {vocab_path} "
+                f"--merge-file {merges_path} "
+                f"--dataset-impl mmap "
+                f"--tokenizer-library megatron "
+                f"--tokenizer-type GPT2BPETokenizer "
+                f"--workers {ncpus} "
+                f"--append-eod "
+            )
+            proc = subprocess.Popen(runcmd.format(flags=flags), shell=True)
+            proc.wait()
+            if rm_extracted:
+                os.remove(extracted_path)
 
 
 if __name__ == "__main__":
